+++ docs/make.jl
@@ -13,6 +13,7 @@ makedocs(;
         edit_link="main",
         assets=[
             "assets/custom.css",
+            "assets/favicon.ico",
             # Search Console (URL-prefix: https://yamanori99.github.io/DistSSHKit.jl/).
             RawHTMLHeadContent(
                 """<meta name="google-site-verification" content="frfWUqaHuYYDmZzSSnBhfguS0Y5YC6zssij5qAot6ww" />""",